@@ -1,57 +1,58 @@
 {
   "name": "ap-react",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "author": "metroninja",
   "private": false,
   "main": "server/app.js",
   "engines": {
-    "node": "6.9.4"
+    "node": "6.10.2"
   },
   "dependencies": {
     "app-module-path": "2.2.0",
     "assets-webpack-plugin": "3.5.1",
-    "autoprefixer": "6.7.0",
-    "babel": "6.5.2",
-    "babel-core": "6.22.1",
-    "babel-loader": "6.2.10",
-    "babel-polyfill": "6.22.0",
-    "babel-preset-es2015": "6.22.0",
-    "babel-preset-react": "6.22.0",
-    "babel-preset-stage-1": "6.22.0",
+    "autoprefixer": "6.7.7",
+    "babel-core": "6.24.1",
+    "babel-loader": "7.0.0",
     "babel-plugin-transform-decorators-legacy": "~1.3.4",
-    "babel-register": "6.22.0",
-    "browserslist": "1.6.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.1",
+    "babel-polyfill": "6.23.0",
+    "babel-preset-env": "1.4.0",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-stage-1": "6.24.1",
+    "babel-register": "6.24.0",
+    "browserslist": "2.1.0",
     "classnames": "2.2.5",
     "compression": "1.6.2",
-    "css-loader": "0.26.1",
-    "express": "4.14.0",
+    "css-loader": "0.28.0",
+    "express": "4.15.2",
     "express-react-views": "0.10.2",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "0.9.0",
-    "history": "4.5.1",
-    "image-webpack-loader": "3.1.0",
+    "extract-text-webpack-plugin": "2.1.0",
+    "file-loader": "0.11.1",
+    "image-webpack-loader": "3.3.0",
     "node-fetch": "1.6.3",
-    "node-sass": "4.3.0",
-    "postcss-loader": "1.2.2",
-    "react": "15.4.2",
-    "react-redux": "5.0.2",
-    "react-dom": "15.4.2",
-    "react-router": "3.0.2",
+    "node-sass": "4.5.2",
+    "postcss-loader": "1.3.3",
+    "prop-types": "15.5.8",
+    "react": "15.5.4",
+    "react-dom": "15.5.4",
+    "react-hot-loader": "3.0.0-beta.6",
+    "react-redux": "5.0.4",
+    "react-router": "4.1.1",
+    "react-router-dom": "4.1.1",
     "redux": "3.6.0",
     "redux-thunk": "2.2.0",
-    "sass-loader": "4.1.1",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.14.0",
-    "webpack-merge": "2.4.0"
+    "sass-loader": "6.0.3",
+    "style-loader": "0.16.1",
+    "url-loader": "0.5.8",
+    "webpack": "2.4.1"
   },
   "devDependencies": {
-    "react-hot-loader": "1.3.1",
-    "webpack-dev-server": "1.16.2"
+    "webpack-dev-server": "2.4.4"
   },
   "scripts": {
     "webpack": "node_modules/.bin/webpack",
     "start": "node server/app.js",
+    "prod": "webpack && node server/app.js",
     "dev": "node server/app.js"
   }
 }
