@@ -6,21 +6,22 @@ const config = require('../config/app.js');
 
 const server = new WebpackDevServer(webpack(webpackConfig), {
   compress: true,
+  host: "0.0.0.0",
   headers: {
     'Access-Control-Allow-Origin': '*',
   },
-  host: '0.0.0.0',
   hot: true,
-  noInfo: false,
   publicPath: webpackConfig.output.publicPath,
-  quiet: false,
   stats: {
     colors: true,
   },
 });
 
-server.listen(config.webpack.port, function() {
+server.listen(config.webpack.port, function(err, result) {
   /*eslint-disable */
-  console.log(`Webpack: PORT=${config.webpack.port}`);
+  if(err) {
+    console.log(`webpack Dev Server error ${err}`);
+  }
+  console.log(`Webpack Dev Server: PORT=${config.webpack.port}`);
   /*eslint-enable */
 });
