@@ -1,10 +1,8 @@
 require('app-module-path').addPath(__dirname);
-
-const AssetsPlugin = require('assets-webpack-plugin');
 const autoprefixer = require('autoprefixer');
+const AssetsPlugin = require('assets-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const browserslist = require('browserslist');
-const merge = require('webpack-merge');
 const webpack = require('webpack');
 
 const config = require('./config/app.js');
@@ -12,117 +10,131 @@ const ASSETS_PATH = config.assets.path;
 const TARGET = process.env.npm_lifecycle_event;
 
 const WEBPACK_DEV_URL = `http://${config.webpack.host}:${config.webpack.port}`;
-
-const sassLoaders = [
-  'css-loader',
-  'postcss-loader',
-  `sass-loader?sourceMap`,
-];
-
 const browsersListConfig = {
   browsers: 'last 2 versions,iOS >= 8,Safari >= 8',
 }
-process.env.BROWSERSLIST = browsersListConfig.browsers;
-
-const common = {
-  entry: {
-    main: "app/client.jsx",
+const sassLoaders = [
+  { loader: 'style-loader' },
+  { loader: 'css-loader', options: { minimize: true, importLoaders: 1 } },
+  {
+    loader: 'postcss-loader',
+    options: {
+      ident: 'postcss',
+      plugins: () => {
+          return [autoprefixer(browsersListConfig)];
+      },
+    }
   },
+  {
+    loader: 'sass-loader',
+    options: {
+      sourceMap: true,
+    }
+  },
+];
+
+isDev = TARGET === 'dev';
+
+const entry = {
+  dev: [
+    'react-hot-loader/patch',
+    `webpack-dev-server/client?${WEBPACK_DEV_URL}`,
+    'webpack/hot/only-dev-server',
+    config.assets.entryScript,
+  ],
+  prod: ['babel-polyfill', config.assets.entryScript],
+}
+const sass = {
+  dev: sassLoaders,
+  prod: ExtractTextPlugin.extract({
+    fallback: 'style-loader',
+    use: sassLoaders.slice(1),
+  }),
+};
+const output = {
+  dev: {
+    filename: '[name].js',
+    path: process.cwd(),
+    publicPath: `${WEBPACK_DEV_URL}${ASSETS_PATH}`,
+  },
+  prod: {
+    filename: '[name]-[hash].js',
+    path: '/src/public',
+  },
+};
+const plugins = {
+  dev: [
+    new webpack.HotModuleReplacementPlugin(),
+    new webpack.NamedModulesPlugin(),
+    new webpack.NoEmitOnErrorsPlugin(),
+    new webpack.DefinePlugin({
+      'process.env': {
+        NODE_ENV: JSON.stringify('development'),
+      },
+    }),
+  ],
+  prod: [
+    new AssetsPlugin({
+      filename: 'asset-manifest.json',
+      path: config.assets.compilePath,
+      prettyPrint: true,
+    }),
+    new ExtractTextPlugin({
+      filename: '[name]-[hash].css',
+    }),
+    new webpack.DefinePlugin({
+      'process.env': {
+        NODE_ENV: JSON.stringify('production'),
+      },
+    }),
+    new webpack.optimize.UglifyJsPlugin({
+      mangle: false,
+    }),
+  ],
+};
+
+module.exports = {
+  cache: true,
+  devtool: isDev ? 'inline-source-map' : 'nosources-source-map',
+  entry: isDev ? entry.dev : entry.prod,
+  output: isDev ? output.dev: output.prod,
+  target: 'web',
   module: {
-    loaders: [{
+    rules: [{
+      test: /\.js.?$/,
+      include: /app/,
+      use: ['babel-loader'],
+    }, {
+      test: /\.scss$/,
+      use: isDev ? sass.dev : sass.prod,
+    }, {
       test: /\.css$/,
-      loader: "style-loader!css-loader"
+      use: ["style-loader", "css-loader"],
     }, {
       test: /\.(eot|svg|ttf|woff|woff2)$/,
-      loader: 'file?name=/fonts/[name].[ext]'
+      use: [{
+          loader: 'file-loader',
+          options: {
+            name: '/fonts/[name].[ext]',
+          },
+      }]
     }, {
       test: /\.(jpe?g|png|gif|svg)$/,
-      loaders: [
-        'file?name=/images/[name].[ext]'
+      use: [
+        {
+          loader: 'file-loader',
+          options: {
+            name: '/images/[name].[ext]',
+          }
+        }
       ]
     }],
   },
-  postcss: [
-    autoprefixer(browsersListConfig),
-  ],
-  progress: true,
-  resolve: {
-    root: [
+  resolve : {
+    modules: [
       process.cwd(),
+      "node_modules"
     ],
   },
+  plugins: isDev ? plugins.dev : plugins.prod,
 };
-if(TARGET === 'dev') {
-  module.exports = merge.smart(common, {
-    cache: true,
-    devTool: 'inline-source-map',
-    entry: [
-      `webpack-dev-server/client?${WEBPACK_DEV_URL}`,
-      'webpack/hot/only-dev-server',
-      config.assets.entryScript,
-    ],
-    module: {
-      loaders: [{
-        test: /\.js.?$/,
-        include: /app/,
-        loaders: ['babel-loader'],
-      }, {
-        test: /\.scss$/,
-        loader: `style-loader!${sassLoaders.join('!')}`,
-      }],
-
-    },
-    output: {
-      filename: '[name].js',
-      path: process.cwd(),
-      publicPath: `${WEBPACK_DEV_URL}${ASSETS_PATH}`,
-    },
-    plugins: [
-      new webpack.HotModuleReplacementPlugin(),
-      new webpack.NoErrorsPlugin(),
-      new webpack.DefinePlugin({
-        'process.env': {
-          NODE_ENV: JSON.stringify('development'),
-        },
-      }),
-    ],
-  });
-} else {
-  module.exports = merge.smart(common, {
-    entry: ['babel-polyfill', config.assets.entryScript],
-    devTool: 'source-map',
-    module: {
-      loaders: [{
-        test: /\.js.?$/,
-        include: /app/,
-        loader: 'babel-loader',
-      }, {
-        test: /\.scss$/,
-        loader: ExtractTextPlugin.extract('style-loader', sassLoaders.join('!')),
-      }],
-    },
-    output: {
-      filename: '[name]-[hash].js',
-      path: config.assets.compilePath,
-    },
-    plugins: [
-      new AssetsPlugin({
-        filename: 'asset-manifest.json',
-        path: 'public',
-        prettyPrint: true,
-      }),
-      new ExtractTextPlugin('[name]-[hash].css'),
-      new webpack.DefinePlugin({
-        'process.env': {
-          NODE_ENV: JSON.stringify('production'),
-        },
-      }),
-      new webpack.optimize.UglifyJsPlugin({
-        compress: {
-          warnings: false,
-        },
-        mangle: false,
-      }),
-    ],
-  });
-}
